Updated Twitter with auth

+++ controller/tweet.js
@@ -1,5 +1,5 @@
 import * as tweetRepository from '../data/tweet.js';
-
+import {} from 'express-async-errors';
 export async function getTweets(req, res, next) {
 	const username = req.query.username;
 	const data = await (username
@@ -10,16 +10,17 @@ export async function getTweets(req, res, next) {
 
 export async function getTweet(req, res, next) {
 	const id = req.params.id;
-	const tweet = await tweetRepository.getAllById(id);
+	const tweet = await tweetRepository.getById(id);
 	if (tweet) {
-		res.status(200).json(data);
+		res.status(200).json(tweet);
 	} else {
 		res.status(404).json({ message: `Tweet id(${id}) not found` });
 	}
 }
 
 export async function createTweet(req, res, next) {
-	const { text, userId } = req.body;
+	const { text } = req.body;
+	const userId = req.userId;
 	const data = await tweetRepository.create(text, userId);
 	res.status(201).json(data);
 }
@@ -27,9 +28,13 @@ export async function createTweet(req, res, next) {
 export async function updateTweet(req, res, next) {
 	const id = req.params.id;
 	const { text } = req.body;
-	const tweet = await tweetRepository.getAllById(id);
+	const tweet = await tweetRepository.getById(id);
 	if (!tweet) {
 		return res.status(404).json({ message: `Tweet id(${id}) not found` });
+	}
+	if (tweet.userId !== req.userId) {
+		return res.sendStatus(403);
+		s;
 	}
 	const updated = await tweetRepository.update(id, text);
 	res.status(200).json(updated);
@@ -37,9 +42,12 @@ export async function updateTweet(req, res, next) {
 
 export async function deleteTweet(req, res, next) {
 	const id = req.params.id;
-	const tweet = await tweetRepository.getAllById(id);
+	const tweet = await tweetRepository.getById(id);
 	if (!tweet) {
 		return res.status(404).json({ message: `Tweet id(${id}) not found` });
+	}
+	if (tweet.userId !== req.userId) {
+		return res.sendStatus(403);
 	}
 	await tweetRepository.remove(id);
 	res.sendStatus(204);
